jansson: force shared library
Summary:
jansson 2.14 switches from autotools to cmake and with
it requires a new option in order to build a shared library.  Without
this cmake option we get tons of linker issues.

The addition of the EXTRA_OEMAKE variable makes no affect on autotools
based versions of the package.

Test Plan: Built halfdome with latest lf-master.

Reviewed By: garnermic

fbshipit-source-id: f61727ed41c5d98fe418c23df6614197cb857844