bletchley: add idevicerestore hacks
Summary:
idevicerestore is used for provisioning and the upstream
code assumes it is ran on a desktop-class system with a LOT of
memory.  Quanta developer has done some hacks in order to get it
to run on a BMC, but we are still working through how we're going
to upstream this work.  Merge as an ugly patch-file now to unblock
PVT provisioning workflows.

Test Plan:
Built and tested boot on QEMU.  Quanta developer has
tested on hardware with real reimaging attempts.

Reviewed By: peterdelevoryas

fbshipit-source-id: 4e47d3ebb9542dd525a381bf34e30b2cf944c5e6