distro: openbmc-fb-lf: clean up openbmc-fb comparison
Summary:
There was an old comment to go back and revisit
the differences between openbmc-fb and openbmc-fb-lf distros.
Most of the differences have been taken care of or are not
relevant.  Clean up the comment and add a DISTROOVERRIDE (as
that was the only remaining functional difference).

Test Plan: Build Bletchley.

Reviewed By: peterdelevoryas

fbshipit-source-id: 22bc18a7b2b6096f813d8bfb6ac8f93ca639afed