fby3: compile fixes in fw-util and libbic for later GCC
Summary:
Later versions of GCC have stricter warnings on unused
parameters and variables; clean these up.  There is also a warning
when using the sscanf "%m" format field, because this is a GNU
extension, so backport the same fix I did recently in common to
use a pre-allocated buffer and "%s".

Test Plan: Built fby3 on lf-master.

Reviewed By: peterdelevoryas

fbshipit-source-id: 9b288d927d0b642f06a5245f9e38ea422238a24b