rackmon2: Make the transaction atomic from an async perspective
Summary:
In the earlier code, where we use actual async streams, each
of the transaction was async and could be independently
cancelled. This could cause some weird behaviors and in
particular cause the rackmon service thread to wait
forever for an "Abandoned" client to send more data.
PS: This could have been fixed with a `await writer.drain()` probably.
But it just helps to have the sync and async code use similar
paths which can make maintenance a lot more easier.

Test Plan:
Load onto wedge400 with a constant reads from the rest's
/api/sys/modbus/read interface and do an overnight soak.

Reviewed By: peterdelevoryas

fbshipit-source-id: 6cfb396d7c918c1101b62fb02ed5c3c592f4c453