No need to remove MCTP header
Summary: The issue T139149658 was introduced by D40656192 when we moved from direct mctp to mctp_d. Direct mctp return includes a MCTP header (4 bytes) in the response buffer, while mctpd removes the MCTP header in the response. Thus, in the attest-util app, we don't need to remove the 4 bytes of MCTP header again.

Test Plan: Not sure the test process. But this is a minor change, please review it carefully.

Reviewed By: amithash

fbshipit-source-id: 93beff8880bd763b18a8846ce15d443d6d734de7