northdome: linux: remove CX7 workaround
Summary:
We want the CX7 workaround enabled for Yv3.5 but not
Northdome currently.  Revamp the bbappend so that the patchfile
is applied based on a MACHINEOVERRIDE.  This causes it to not
apply to halfdome.

Test Plan:
Run `bitbake -ne` on Northdome and fby35.

  Northdome:
```
SRC_URI="file://linux-aspeed-5.15 file://ubifs.scc \
    file://ubifs.cfg file://defconfig file://fby35.cfg"
```
  fby35:
```
SRC_URI="file://linux-aspeed-5.15 file://ubifs.scc \
    file://ubifs.cfg file://defconfig file://fby35.cfg \
    file://0001-workaround-for-CX7-set-MAC-Affinity-issue.patch"
```

Reviewed By: peterdelevoryas

fbshipit-source-id: e08ed584c208f87a27f114f4b25b1d569364d5c7