Fix IPv6 Addresses Returned
Summary: Redfish is a REST API for OpenBMC, we have an endpoint which returns the ipv6 address of the bmc. This commit fixes the code to read from /proc/net instead of parsing output of the `ip a show` command. resulting in faster and more accurate response.

Test Plan: {F819701699} Canaried on a wedge-100 test machine, and validated the output.

Reviewed By: kawmarco

fbshipit-source-id: 471b515a52b494d15f7a617d3a2353dc53ef5a89