fw-util: avoid sig-unsafe calls in signal handler
Summary:
In signal handlers there is a very very small set of functions which
can safely be called.  Neither `cout` nor `syslog` are in that set.

We have ran into cases where `fw-util` processes have hung after tooling
attempts to run `fw-util all --version` and experiencing a timeout.  I
observed one system with dozens of hung `fw-util` processes.  When
connecting to them with `strace`, they were hung in a `futex` call, and
when attaching with `gdb` the stack was corrupted.  My best theory is
that we get a signal in the middle of something complex like a memory
allocation and then the `syslog` (which also calls malloc) corrupts
the process.

Simplify all code in the signal handler to simply save away the signo
and then move all logging over to the normal part of the application.

Test Plan:
Forced a SIGHUP in QEMU and observed the expected behavior.

cout -
```
Aborted action due to signal
```

syslog -
```
 2022 Dec 14 17:29:12 bmc-oob. user.debug fby3-fcabf582d7-dirty: fw-util: fw-util: previously received signal=1
 2022 Dec 14 17:29:12 bmc-oob. user.debug fby3-fcabf582d7-dirty: fw-util: fw-util: Terminate request handled
 ```

Reviewed By: peterdelevoryas

Differential Revision: D42055287

fbshipit-source-id: 286e4d557ec10b1dc4545071367fc52f20cb52fe