Fix test_fw upgrade json dryrun
Summary:
test_fw upgrade json dryrun was failing because there are lots
of several fpd that was removed in the json. Also, some changes in the
expected command was made as well causing the test to fail. The change
below fix those firmware issue and ensure that they all pass

Test Plan: see the following paste P579527488

Reviewed By: binhuang00

Differential Revision: D42090388

fbshipit-source-id: 51edaa2074d0e5a471b8a0eb5a4eba73cf38d3b9