rackmon2: Allow configuring python-path
Summary:
The validator scripts are called by whatever python3 the path
has. Unfortunately some (my) setups have multiple python
installations and the one in the env is not the one with
the required modules installed (python3-jsonschema).

Thus, allow the user to configure the default python build
and hence allowing the user to provide the path of the system
default python. The default (used by yocto) will continue
to use the one present in the env, so this should be a no-op.

Test Plan:
Build all the platforms.
Try:
```
meson -Dsyslog=false -Dpython-path=/usr/bin/python3 /tmp/builddir
ninja -C /tmp/builddir
```

Reviewed By: garnermic

Differential Revision: D42902906

fbshipit-source-id: c84ce3e3755645a5212ee7bcb74bebdf8c126995