rackmon2: Cleanup redundant code
Summary:
Now that we no longer support the "legacy" API, there
is really no need to continue to do the specific if/else
for JSON parsing. Now consider it a generic error. It
cleans up the code and makes it easier to read.

Test Plan: Build wedge400 and ensure all unit tests pass

Reviewed By: williamspatrick

Differential Revision: D42971361

fbshipit-source-id: b03de27dfd5b02b81ca7d2fad7ddaf3364addd70