Elbert: Fix switchcard DPM dump utility (#287)
Summary:
The DPM dump logic for Elbert is currently not decoding the fault registers from the DPM correctly. The result is that the information in the dpm_log file is missing the fault bytes.

This change fixes the dump logic so that these missing bytes are included.

Testing performed:

Power cycled the system by removing and re-applying external power. The last line of the /mnt/data1/log/dpm_log file is below. There are now 10 bytes for each fault log. Previously, only the byte length was logged.

```
{'Date':'2018-03-09 12:34:57','Source':'SWITCHCARD','Log':'0x01 0x02 0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x02 0x02 0x00 0x00 0x00 0x00','detail_0':{'m':'0x02 0xb5 0x47 0xe4 0x48 0x80 0x44 0xbf 0x00 0x00','e':'0x12'},'detail_1':{'m':'0x02 0xb5 0x47 0xfc 0x8d 0x00 0x44 0xbf 0xed 0x25','e':'0x12'},'detail_2':{'m':'0x02 0xb5 0x47 0xfc 0x8d 0x80 0x44 0xbf 0x0d 0x25','e':'0x12'}}
```

Reviewed By: rootkeeper

Differential Revision: D43669511

Pulled By: joancaneus

fbshipit-source-id: c76160c5be5af66c82e46e9e63c0112247eeacf9