bletchley: libimobiledevice-glue: add http proxy support
Summary:
We leverage some software from 'libimobiledevice' in order
to perform low-level reprovisioning of the Mac Minis in the Bletchley
chassis.  This operation performs a "DFU wipe" and re-installs the OS.
As part of the OS reinstall, the Mac expects to be able to reach back
to Apple and libimobiledevice already has some amount of reverse-proxy
support.

Unfortunately(?), the BMC is not able to directly access the internet
in production networks, so this DFU/Reinstall is unsuccessful.  Nima
was able to modify libimobiledevice-glue to add support for additional
proxy environment variables, which tunnels the Mac network operations
through an acceptable proxy.  This is currently applied as a patch
until we are able to fully upstream this support.

Test Plan:
Built test image and Nima was able to successfully
reprovision the Mac Mini.

Reviewed By: garnermic

Differential Revision: D43878886

fbshipit-source-id: dc6ef3e1d27cd453e3ee27e5bb238b7e4ba0f84f