- Automate feature build for common platform features
Summary:
This patch introduces the ability to automatically build the feature doc
for a platform if it doesn't exist using the feature files in the common
directory and a default features files as pointed to in the platform's
YAML file.  The change in the platform YAML file is to add a features
section as shown below which indicates the name of the default file to
use to build the features list.

	feature_list:
	  - default-meta-multihost-features

In this example, the DEFAULT-META-MULTIHOST-FEATURES.json file would be used.
The code will search both the common directory and the platform directory
to find this file if it exists (the platform overrides the common one if
the file is found in both locations)

The build-machine-spec also creates a docx file using pandoc if it is
installed along with the output file as indicated by the user.  The test
plan has links to both docs created from the default file for the
Great Lakes system.

In a future patch set, the ability to add platform specific features
in addition to the common ones will be added.

Test Plan:
The test file shows the output from the tool in markdown.  The test.docx was created from the test file.

The output used to generate this file using the tool is (executed from the meta-facebook/spec directory):

./tools/build-machine-spec -m greatlakes -o test

The expectation is that all the diffs labeled as [SPEC][COMMON] have been landed to provide the content information.  If they are not found, the feature file will show the "**Missing features md file**" error for each feature without a corresponding feature file available.

{F899238209} {F899238295}

Console output from the above command:
[garnermic@devvm7271.ftw0 ~/local/openbmc/meta-facebook/spec (miscellaneous)]$ ./tools/build-machine-spec -m greatlakes -o test
Writing section overview
Feature file successfully loaded.
Writing section features
Finding feature file ... DEFAULT-META-MULTIHOST-FEATURES.json
Feature file successfully created.
Output docx file test.docx successfully generated
Feature file generation completed.

Reviewed By: williamspatrick

Differential Revision: D43411714

fbshipit-source-id: 827333cb12d46ba3d4b1b3f25ee95026138a7da3