Add redfish dump utility to repo
Summary:
The redfish dump utility enables dumping the complete Redfish tree
for a target server starting at the /redfish/v1 API point.  The output
is a JSON file which stores all the Redfish data retrieved from the
target server.

Examples of target servers for both Meta and LF are included in the code.
The code must be manually modified at the moment for the target system

Test Plan: See output generated in exampled rfData.json file

Reviewed By: amithash

Differential Revision: D43992490

fbshipit-source-id: a7d166d4da8c016af10b5be83b808f117aeba4a4