openbmc: clean up meson.build `find_library` calls
Summary:
Meson has a call cc.find_library(...) which uses the compiler to find the library and create a meson dependency from that.  Many of these were originally created because the package providing the library didn't provide a corresponding pkg-config file (foo.pc).  Since then, many of the packages have been converted to Meson themselves and should be switched to pkg-config based dependency instead of library-based dependency.

Typically the solution is cc.find_library('foo') becomes dependency('foo').  The complete solution may depend on the .pc file supplied by the dependent package.  You can find these in the tmp/work/... tree of the Yocto build.

Go through the find_library calls in the above paste.  Determine which of the 3 classes it fits in:

Upstream packages that do not provide a .pc file.
    - These are rare, but can be left alone.
Local packages that have converted to Meson now.
    - Upload diff(s) to change the find_library call to the dependency call.
Local packages that have not switched to Meson.
    - Create a task for each of these and assign back to Patrick.

Test Plan: CIT build passes.

Reviewed By: williamspatrick

Differential Revision: D44364278

fbshipit-source-id: 5df91e343e055a7de325c96675bb26040f7008d0