common: adjust ensurepip removal
Summary:
With D19265615 we removed the ensurepip module from python3 in the
images because we don't actually use pip on-device even though some
python modules RDEPEND on it.  The latest Yocto moved the ensurepip
into its own package and isn't happy if we delete the whole package.
Adjust the removal so that we're only deleting the large pip binaries
instead of every file in the package.

Test Plan:
Build rocko and lf-master machines and ensure the large
pip binaries are still removed.

Reviewed By: jagpalgill

Differential Revision: D45946750

fbshipit-source-id: 46781e2282aefb4585c350ca0766dd0b5e784cce