linux-aspeed: disable GCC plugins on 5.x kernels
Summary:
The 5.x series kernels have a GCC plugin which no longer
compiles on GCC-13.  This is used in development to spot certain
kernel-specific bugs, but this shouldn't be necessary for us.  Disable
the plugins in order to be able to compile with the latest lf-master
refresh.

Test Plan: Successfully built with GCC-13.

Reviewed By: jagpalgill

Differential Revision: D46766116

fbshipit-source-id: f75a7d6c489d6c62dde77cab3e2322c876aacedf