libnm: rename to nodemanager
Summary:
libnm is used for interacting with the Node Manager in Intel
processors, but is a short enough name to be ambiguous.  The latest
Yocto refresh adds `libnm` support to the networkmanager repository
which causes confusion as to which package is desired.  Rename our
libnm to libnodemanager to avoid the confusion and ensure the build can
pass.

Test Plan: Bitbake warning about PREFERRED_PROVIDER is gone.

Reviewed By: jagpalgill

Differential Revision: D46772461

fbshipit-source-id: fb9bc24005a1b5e3dcae28f28be9c5f74e780d72