spec: adjust feature YAML format
Summary:
Change the feature YAML format so that the feature title
is a dictionary key, rather than an array of dictionaries with
"title" as a key.  This is, hopefully, more intuitive and also makes it
easier to merge status between two different dictionaries: one for
the feature definition, which is common, and one for the feature
status, which is machine-specific.

Test Plan: Regenerate the greatlakes spec document.

Reviewed By: amithash

Differential Revision: D47317864

fbshipit-source-id: 6e7b3425bd1b0ef3d2a7a205d3a57446969747a7