rackmon: Handle connection close condition.
Summary:
recv() manpage clearly states
```
When a stream socket peer has performed an orderly shutdown, the
return value will be 0 (the traditional "end-of-file" return).
```
Thus, if the remote end closes the socket we will end up with
an infinite loop waiting for data.

Test Plan:
Custom hack a test client which starts an connection and
immediately terminate. Check that the service is no longer
viable and accepting connections and cannot be terminated.

After the fix the test client no longer puts the service
in the bad state.

Reviewed By: williamspatrick

Differential Revision: D47852141

fbshipit-source-id: eee752c2d26f4fb4314fd4a6d3c9df4af0af645a