meta-facebook: move serfmon service from montblanc to fboss-lite
Summary:
Move "console-autodiscovery" service from montblanc to fboss-lite layer,
and rename the service to serfmon.

NOTE: the package is removed from montblanc image, because we are not
sure if the service is really needed. If later we decide to include the
package, it should be handled in fboss-lite layer.

Test Plan:
1) "bitbake montblanc-image" succeeded.
2) include "serfmon" package in fboss-lite-image.inc and confirmed the
   service is installed in montblanc image properly.

Reviewed By: Scott8440

Differential Revision: D48415572

fbshipit-source-id: b1c3d52b1d875b63cdcc5366ea4a9003e5c35133