tpm2: remove python3-tpm2-pytss dependency
Summary:
The latest Yocto added a python3 dependency as part of
the TPM2 package group.  Remove it so that we don't end up with
as big of a dependency chain.  Currently, having this in place
is causing a hang in the `do_install` step of the package on
Bletchley.

Test Plan: Sandcastle successful for Bletchley.

Reviewed By: jagpalgill

Differential Revision: D49154313

fbshipit-source-id: e50572375ecf1a0c6bdbf2569290c7e212b73419