Generate eeprom
Summary:
The eeprom.json was originally hardcoded into the fboss-lite reference layer, but we've found that it needs to be customizable for each platform.
This is a super simple addition to the openbmc config which is basically copied into eeprom.json.
No change to the bitbake file since it does not need to change across platforms

Test Plan:
```
{
    "platform_name": "wedge400-test",
    "eeproms": [
        {
            "name": "bmc_eeprom",
            "address": "/sys/bus/i2c/devices/i2c-8/8-0051/eeprom"
        },
        {
            "name": "chassis_eeprom",
            "address": "/sys/bus/i2c/devices/i2c-6/6-0053/eeprom"
        }
    ]
}
```
Output into `<layer>/recipes-utils/wedge-eeprom/files/lib/eeprom.json`

Reviewed By: tao-ren

Differential Revision: D49204730

fbshipit-source-id: 45334954831b1266b0692313896d5d58162ffcfb