common: moved showtech.sh from fboss-lite to fbmc-snapshot package
Summary:
# Description

A new package fbmc-snapshot is added in common recipes-utils and moved the showtech.sh script in this package for common use and included in Fuji image and Montblanc image.

# Motivation

The ultimate goal is to make a new package "**fbmc-snapshot**" in the common layer which can execute the common showtech.sh" script along with the fboss-lite layer specific and platform specific showtech plugins.
The "**showtech.sh**" is a script which will list the device info like hostname, uptime, BMC version, Aspeed chip version, BMC board revision, etc,. This package "fbmc-snapshot" is to make the existing showtech.sh script common for all platforms. The layer(fboss-lite) specific scripts and platform specific scripts can be added and installed in the device path `/etc/showtech/rules/` using bbappend file of this fbmc-snapshot package. And the common showtech.sh will execute all the scripts which are placed in `/etc/showtech/rules/` as a part of showtech.

The plugins can be added with the naming conventions as follows,
File name starts with
1. 0-99 --> Common scripts.
2. 100-200 --> Fboss-lite
3. 900-999 --> Platform specific.

So that we can use this show tech script to get all the info regarding the platform.

X-link: https://github.com/facebookexternal/openbmc.celestica/pull/1647

Test Plan:
1. Verify the presence of showtech.sh in `/usr/local/bin/`
2. Execute the "showtech.sh" and verify the output.

**Unit test log:**
[mp3_showtech_output_log.txt](https://github.com/facebookexternal/openbmc.celestica/files/12856306/mp3_showtech_output_log.txt)

Reviewed By: mikechoifb

Differential Revision: D50119975

fbshipit-source-id: 9b38eaa52305ae481727b686661c1c131b29177e