openbmc: Reformat data0 if its size has increased
Summary:
As a preparation for the next diff which increases the data0
partition size, this diff updates the `mount_data0.sh` script that is
run at boot to check if the partition size is larger than the UBI volume
size (which is a sign that the partition has been enlarged), and if so,
recreates the volume at the new size and formats it.

Test Plan:
In the next diff. Also:
```
/dev/ubi0_0 has size 7979776, but device /dev/ubi0 has size 66977792. Recreating volume...
ubiformat: mtd4 (nor), size 67108864 bytes (64.0 MiB), 1024 eraseblocks of 65536 bytes (64.0 KiB), min. I/O size 1 bytes
libscan: scanning eraseblock 1023 -- 100 % complete
ubiformat: 1024 eraseblocks have valid erase counter, mean value is 0
ubiformat: formatting eraseblock 1023 -- 100 % complete
```

Reviewed By: tao-ren

Differential Revision: D50813286

fbshipit-source-id: 501c8b260a275bd806a98eaaad86abdd65e5d452