sensord: Allow monitoring intervals less than default
Summary:
As title, this should be a no-op since pal_get_sensor_poll_interval
returns 2s as default for most sensors but opens us up to
a smaller monitoring time if required.

The simplest change would be to change the polling time to 1s by default
but on most platforms that would be an unnecessary wake to sleep again.
Hence add extra logic before starting the monitoring loop to get the
poll interval which now is no longer a compile time constant.

Test Plan:
Build grandteton and check /var/log/messages that all FRUs are using
the default polling interval:
```
      sensord: Using polling interval 2 for FRU: 1
      sensord: Using polling interval 2 for FRU: 2
      sensord: Using polling interval 2 for FRU: 3
      sensord: Using polling interval 2 for FRU: 8
      sensord: Using polling interval 2 for FRU: 9
      sensord: Using polling interval 2 for FRU: 10
      sensord: Using polling interval 2 for FRU: 11
      sensord: Using polling interval 2 for FRU: 12
```
define the polling interval to switch some sensors to use 1s as the
polling interval. Ensure sensord switches appropriately
```
      sensord: Using polling interval 2 for FRU: 1
      sensord: Using polling interval 2 for FRU: 2
      sensord: Using polling interval 2 for FRU: 3
      sensord: Using polling interval 2 for FRU: 8
      sensord: Using polling interval 1 for FRU: 9
      sensord: Using polling interval 1 for FRU: 10
      sensord: Using polling interval 2 for FRU: 11
      sensord: Using polling interval 2 for FRU: 12
```

Reviewed By: williamspatrick, dbxfb

Differential Revision: D50618293

fbshipit-source-id: 96e711e1181bc7b9a3c8da6200d801bee3f22706