morgan800cc: openbmc-utils: add setup-gpio.sh
Summary:
# Description

This change is to setup and export the GPIOZ7. It is used for toggling the mux used to access the device on spi1 which is setup in https://github.com/facebookexternal/openbmc.cisco/issues/82.

- The config registers needed to be set in order for the GPIOZ7 to be exported.
- The direction is setup as out irrespective of the default value.

This script is not included in openbmc-utils%.bbappend because there will be a conflict with the subsequent PR that introduces the script. Will included along with that.

# Motivation

This change is to setup and export the GPIOZ7 which is used to toggle the mux to access the BIOS flash. This is required for the BIOS recovery.

X-link: https://github.com/facebookexternal/openbmc.cisco/pull/83

Test Plan:
1. "bitbake morgan800cc-image" succeeded.
2. One error in journalctl with respect to setup-gpio.sh (after including them in the image). This error is because the /var/ directory was not created when this service was run. Will be fixing this as a separate PR.
```
root@bmc:~# journalctl -u setup_gpio
Sep 20 04:23:31 bmc setup-gpio.sh[97]: /usr/local/bin/board-utils.sh: line 48: /var/lock/lockfile_for_board_util: No such file or directory
Sep 20 04:23:31 bmc setup-gpio.sh[97]: exporting gpio (aspeed-gpio, GPIOZ7), shadow=BMC_SPI1_MUX
Sep 20 04:23:32 bmc systemd[1]: Finished Setup GPIOs in OpenBMC.
```
3. Able to access GPIOZ7 via provided gpio utilities.
4. Able to use flip the mux connected to this GPIO to mount the bios flash mtd partition.

Reviewed By: somasun

Differential Revision: D52097381

Pulled By: tao-ren

fbshipit-source-id: 7897e4c98899f7f9e39a71c4b2ab6f8ad1d94a2a