flake8: disable warnings
Summary:
With the latest flake8, we are seeing a large number of new warnings
of these two classes:

```
/usr/local/fbpackages/rest-api/test_redfish_bios_firmware_dumps.py:320:21: W503 line break before binary operator
/usr/local/fbpackages/rest-api/test_rest_modbus_cmd.py:3:1: F401 'typing.Any' imported but unused
```

For the W503, this is a style fight between flake8 and black, so we
can disable it in favor of black.

For F401, while this is not a great option to disable, the newer version
of flake8 disabled support for the old-style `# type: Any` style
comments and expect you to use the modern `var: Any =` syntax.
Unfortunately that syntax is not supported on Python 3.5 that we still
need to support for Rocko.  Add 'noqa' statements for these.

Test Plan: Newer flake8 passes in ptest.

Reviewed By: kawmarco

Differential Revision: D53192114

fbshipit-source-id: 7d24e4a5f7ec87fd39aa26dc7fc42b43aa2de8a6