mfgtool: inventory: add empty interfaces
Summary:
DBus interfaces can be empty and in the case of inventory
might be keys related to the type of inventory item.  Ensure that
the inventory output includes empty interfaces.

Test Plan:
Ran on a Yv4 system to ensure no regressions but there are
currently no examples of empty interfaces in the inventory hierarchy.

Reviewed By: wangx6f

Differential Revision: D54665806

fbshipit-source-id: 2e719c3343f26db63e8d3b4a4fca47e714b011d6