common: fbmc-snapshot: add "dmesg" and "/var/log/messages" in showtech.sh script
Summary:
# Description

As a part of fbmc_snapshot enhancement, **000_collect_log_messages.sh** is added as a common plugin in a common layer in which logs of `dmesg` and `/var/log/messages` are added. This script is installed in  `/etc/showtech/rules/` path so that it will be executed by showtech.sh script. We could get more details about the device.
The plugins naming conventions is altered in the README file of this fbmc-snapshot package.

# Motivation

To get more info about the device by adding the **dmesg** log and **/var/log/messages** content as a part of the showtech output.

X-link: https://github.com/facebookexternal/openbmc.celestica/pull/1666

Test Plan:
1. Build the image for Fuji and Janga.
1. Verify the presence of **000_collect_log_messages.sh** script in `/etc/showtech/rules/`
1. Verify the output of showtech by executing the command showtech.sh in the device(Fuji, Janga)

**Unit test logs**
[Janga_qemu_show_tech_output.txt](https://github.com/facebookexternal/openbmc.celestica/files/13315884/Janga_qemu_show_tech_output.txt)
[fuji_show_tech_output_1.txt](https://github.com/facebookexternal/openbmc.celestica/files/13315885/fuji_show_tech_output_1.txt)

Reviewed By: mikechoifb

Differential Revision: D55658689

Pulled By: alandau

fbshipit-source-id: d4f3ca0b3779a500e69cbf417b56c7c84f0b5247