Fixing firmware_ext testing failing
Summary: Firmware_ext was constantly failing on several conveyor nodes. We found issue with the sha1sum parser which we fixed. However, the most complicated issue where the fact that accessing units were failing on sandcastle server but not locally. We cleaned up some codes, increase time out, add some retry within a theshold to improve the test signals. We tried paramiko but it was performing worst on sandcastle for some unknown reason. All our tests has worked as expected, we will keep some debug statement in this diff so that we can monitoring closely in the next two months and debug if needed. We will remove the dbg statement after letting this run on the conveyor for a good 2 months.

Test Plan:
Fuji:
https://www.internalfb.com/intern/sandcastle/job/22517999411485039/

Wedge400:
https://www.internalfb.com/intern/sandcastle/job/1277921679/

yamp:
https://www.internalfb.com/intern/sandcastle/job/22517999411742676/

Minipack:
https://www.internalfb.com/intern/sandcastle/job/31525198668753485/

elbert:
https://www.internalfb.com/intern/sandcastle/job/36028798297030188/

Reviewed By: cjcon90

Differential Revision: D57091670

fbshipit-source-id: 6c0d70ff4e1bcc1bdbc25b13caa4a5cb7ae36907