mfgtool: power-control: display exceptions
Summary:
When the power-control experiences any exception it just
displays `failed`.  This is not very useful for debugging what the
underlying issue might be.  Modify the `catch` clause to display the
exception.

Ensure that all of the exceptions created by the application itself
have real exception types and error strings rather than throwing an
empty `std::exception`.

Test Plan:
Compiled and ran on x86.  Verified that a default exception
message is created and displayed.

```
$ ./builddir/mfg-tool power-control --action=off --scope=runtime --position=1
<6> Attempting runtime:off on 1.
<4> Can't find /xyz/openbmc_project/state/host1
<6> Trying to escalate to scope=standby
<4> Can't find /xyz/openbmc_project/state/chassis1
<3> Cannot find chassis to execute operation against; giving up.
<3> Caught exception: missing chassis
"failed"
```

Reviewed By: ifel

Differential Revision: D58169074

fbshipit-source-id: 5fadc2db5ea9eed27b8b1089e7585f9a5c712ce8