tiogapass: disable P12V_STBY re-disable
Summary:
Per S425256, we are seeing a large number of Tiogapass
systems ending up effectively bricked.  The P12V_STBY is being
disabled due to a critical condition and then it is continuing to
be disabled over and over, effectively bricking the machine.

This code was introduced after the last long-running BMC release
on Tiogapass without any obvious system damage having taken place.
Disable the "re-disable P12V_STBY" for the time being so we can get
more data on what is happening.

Test Plan: #doitlive

Reviewed By: doranand

Differential Revision: D58782779

fbshipit-source-id: df2a40f8b6b643e18c04b01fb4e2e14341c0b1b5