CPU utilization test: Update thresholds
Summary:
The test still fails sometimes. This is inherent, since a lot
of things happen on the box and its CPU is pretty slow.

Previously, we required that in at least 2 out of 4 periods (of 5s each)
the CPU utilization is below a threshold (mostly 60% but some platforms
override it).

This diff changes this to 3 out of 10 periods. While this may seem low,
it proves that we don't have a regression that pegs the CPU at 100% on
the one hand, and allows 35s (7 periods * 5s) for the noise to settle.

Test Plan: CIT and watch the flaky tests dashboard

Reviewed By: doranand

Differential Revision:
D61829581

Privacy Context Container: L1188897

fbshipit-source-id: 360926cd524004b26a853751aa827cff546aaedf