minipack: Fix sensors going unhealthy
Summary:
Due to the unit type set to `forking`, systemd expected the
`ipmbd` process to fork into the background within 90s. Since this
didn't happen, it killed and restarted it. As a result, the service was
restarted every 90s.

Sensor readings, whether using `sensor-util` or via redfish, fetch some
of the sensors from `ipmbd`, either directly, or a cached value.
Therefore, if the service was restarted at the time when a sensor reading
was requested (not cached), the sensor was shown as `N/A`.

Test Plan:
```
curl 'https://fboss308930394-oob.snc1.facebook.com/redfish/v1/Chassis/1/Sensors?$expand=1' | jq | grep Health | grep -v OK
```
in a loop.

Before this diff, several times an hour it showed:
```
        "Health": "Critical"
```
which is the same as `service-util`'s: `P1V05_MIX_VOLT (0x8E) : NA | (na)`

With this diff, `Health` is always `OK`.

Reviewed By: tao-ren

Differential Revision:
D61891299

Privacy Context Container: L1188897

fbshipit-source-id: 870f96cd25b1ff8a3021bf1e9efe3c1bb3345e7f