elbert: Retry BIOS flash `verify` on error
Summary:
Quoting Arista, "we had some flakiness in the spi reading which
we added retries in BMC software for". The added retries (D45877123) are the
`retry_command` in this file, but retries were missed for the `verify`
subcommand.

This diff adds retries for `verify`.

Test Plan: Eyeballing

Reviewed By: Scott8440

Differential Revision:
D62525122

Privacy Context Container: L1188897

fbshipit-source-id: c6d4f2bbf1ac3ea07bf8da92a4b25b3ae5971d8a