rackmon2: Let device store a reference to the register map
Summary:
This solves a bunch of problems where the device would
need access to a reference to the register maps in the
DB.

Potentially, in future when there is ambiguity of which
register map a device belongs to which is not fully
resolved with the address, we would need this available.

Test Plan: All unit tests pass.

Reviewed By: doranand

Differential Revision:
D63417446

Privacy Context Container: L1188897

fbshipit-source-id: 901b282e21164c15a40347d300d698533d9cb85b