rackmon2: Add a find API to RegisterMapDatabase
Summary:
This adds a find() API to the database class which
returns a const reference to the entires it contains.
This allows a user to request for all register maps
which match a given address instead of getting just
the one.

Keep the at() method and make it call the find()
interface.

Test Plan: All unit tests continue to pass

Reviewed By: doranand

Differential Revision: D63417448

fbshipit-source-id: 08da24d3224f125641a1c376143073fcae222d46